component and styled-component in different files
